+++ src/argilla_server/apis/v1/handlers/datasets/__init__.py
@@ -15,8 +15,10 @@
 from fastapi import APIRouter
 
 from argilla_server.apis.v1.handlers.datasets.datasets import router as datasets_router
+from argilla_server.apis.v1.handlers.datasets.questions import router as questions_router
 from argilla_server.apis.v1.handlers.datasets.records import router as records_router
 
 router = APIRouter(tags=["datasets"])
 router.include_router(datasets_router)
+router.include_router(questions_router)
 router.include_router(records_router)
