+++ src/argilla_server/enums.py
@@ -61,6 +61,7 @@ class QuestionType(str, Enum):
     label_selection = "label_selection"
     multi_label_selection = "multi_label_selection"
     ranking = "ranking"
+    span = "span"
 
 
 class MetadataPropertyType(str, Enum):
